chore: cargo fmt + clippy cleanup across the Phase 7C foundation
- detect::stale_temp: EXEC_EXTS.iter().any(|x| *x == e.as_str())
  → EXEC_EXTS.contains(&e.as_str()) (clippy::useless_iter).
- walker::bomb_guard: 1 * 1024 * 1024 * 1024 → 1024 * 1024 * 1024
  (clippy::no_effect_replace_with_default).
- hasher_mmap test:  cast removed since  was already
  u32 (clippy::cast_lossless / unnecessary_cast).
- walker::dev_excludes test: pack.iter().copied().collect() →
  pack.to_vec() (clippy::iter_cloned_collect).
- cargo fmt auto-applied across every new TASK-202..233 module
  touched by the foundation commits.

Validation: cargo clippy --workspace --all-targets -- -D warnings
clean; cargo test -p mythkernel: 440 unit + 2 e2e pass (was 370
pre-7C, +70 tests across the 15 new modules); cargo fmt --check
clean.

+++ crates/mythkernel/src/hasher_mmap.rs
@@ -119,7 +119,7 @@ mod tests {
     fn mmap_hash_matches_slice_hash() {
         let td = tempfile::tempdir().unwrap();
         let p = td.path().join("payload.bin");
-        let bytes: Vec<u8> = (0u32..10_000).flat_map(|i| (i as u32).to_le_bytes()).collect();
+        let bytes: Vec<u8> = (0u32..10_000).flat_map(|i| i.to_le_bytes()).collect();
         {
             let mut f = std::fs::File::create(&p).unwrap();
             f.write_all(&bytes).unwrap();
